@@ -6,32 +6,178 @@ namespace BGalati\MonologSentryHandler;
 
 use Monolog\Formatter\LineFormatter;
 use Monolog\Handler\AbstractProcessingHandler;
+use Monolog\Level;
 use Monolog\Logger;
+use Monolog\LogRecord;
 use Sentry\Breadcrumb;
-use Sentry\FlushableClientInterface;
+use Sentry\Event as SentryEvent;
+use Sentry\EventHint;
 use Sentry\Severity;
 use Sentry\State\HubInterface;
 use Sentry\State\Scope;
 
+/**
+ * Compatibility layer copied and completed from Sentry SDK.
+ *
+ * @see \Sentry\Monolog\CompatibilityProcessingHandlerTrait
+ */
+if (Logger::API >= 3) { // @phpstan-ignore-line - Comparison operation ">=" between 3 and 3 is always true.
+    /**
+     * Logic which is used if monolog >= 3 is installed.
+     *
+     * @internal
+     */
+    trait CompatibilityProcessingHandlerTrait
+    {
+        /**
+         * @param array<string, mixed>|LogRecord $record
+         */
+        abstract protected function doWrite($record): void;
+
+        /**
+         * {@inheritdoc}
+         */
+        protected function write(LogRecord $record): void
+        {
+            $this->doWrite($record);
+        }
+
+        /**
+         * Translates the Monolog level into the Sentry severity.
+         */
+        private static function getSeverityFromLevel(int $level): Severity
+        {
+            $level = Level::from($level);
+
+            switch ($level) {
+                case Level::Debug:
+                    return Severity::debug();
+                case Level::Warning:
+                    return Severity::warning();
+                case Level::Error:
+                    return Severity::error();
+                case Level::Critical:
+                case Level::Alert:
+                case Level::Emergency:
+                    return Severity::fatal();
+                case Level::Info:
+                case Level::Notice:
+                default:
+                    return Severity::info();
+            }
+        }
+
+        /**
+         * Translates the Monolog level into the Sentry breadcrumb level.
+         *
+         * @param int $level The Monolog log level
+         */
+        private function getBreadcrumbLevelFromLevel(int $level): string
+        {
+            $level = Level::from($level);
+
+            switch ($level) {
+                case Level::Debug:
+                    return Breadcrumb::LEVEL_DEBUG;
+                case Level::Info:
+                case Level::Notice:
+                    return Breadcrumb::LEVEL_INFO;
+                case Level::Warning:
+                    return Breadcrumb::LEVEL_WARNING;
+                case Level::Error:
+                    return Breadcrumb::LEVEL_ERROR;
+                default:
+                    return Breadcrumb::LEVEL_FATAL;
+            }
+        }
+    }
+} else { // @phpstan-ignore-line - Else branch is unreachable because previous condition is always true.
+    /**
+     * Logic which is used if monolog < 3 is installed.
+     *
+     * @internal
+     */
+    trait CompatibilityProcessingHandlerTrait
+    {
+        /**
+         * @param array<string, mixed>|LogRecord $record
+         */
+        abstract protected function doWrite($record): void;
+
+        /**
+         * {@inheritdoc}
+         */
+        protected function write(array $record): void
+        {
+            $this->doWrite($record);
+        }
+
+        /**
+         * Translates the Monolog level into the Sentry severity.
+         *
+         * @param Logger::DEBUG|Logger::INFO|Logger::NOTICE|Logger::WARNING|Logger::ERROR|Logger::CRITICAL|Logger::ALERT|Logger::EMERGENCY $level The Monolog log level
+         */
+        private static function getSeverityFromLevel(int $level): Severity
+        {
+            switch ($level) {
+                case Logger::DEBUG:
+                    return Severity::debug();
+                case Logger::WARNING:
+                    return Severity::warning();
+                case Logger::ERROR:
+                    return Severity::error();
+                case Logger::CRITICAL:
+                case Logger::ALERT:
+                case Logger::EMERGENCY:
+                    return Severity::fatal();
+                case Logger::INFO:
+                case Logger::NOTICE:
+                default:
+                    return Severity::info();
+            }
+        }
+
+        /**
+         * Translates the Monolog level into the Sentry breadcrumb level.
+         *
+         * @param int $level The Monolog log level
+         */
+        private function getBreadcrumbLevelFromLevel(int $level): string
+        {
+            switch ($level) {
+                case Logger::DEBUG:
+                    return Breadcrumb::LEVEL_DEBUG;
+                case Logger::INFO:
+                case Logger::NOTICE:
+                    return Breadcrumb::LEVEL_INFO;
+                case Logger::WARNING:
+                    return Breadcrumb::LEVEL_WARNING;
+                case Logger::ERROR:
+                    return Breadcrumb::LEVEL_ERROR;
+                default:
+                    return Breadcrumb::LEVEL_FATAL;
+            }
+        }
+    }
+}
+
 class SentryHandler extends AbstractProcessingHandler
 {
-    /**
-     * @var HubInterface
-     */
-    protected $hub;
+    use CompatibilityProcessingHandlerTrait;
 
-    /**
-     * @var array
-     */
-    private $breadcrumbsBuffer = [];
+    protected HubInterface $hub;
+    private array $breadcrumbsBuffer = [];
 
     /**
      * @param HubInterface $hub    The sentry hub used to send event to Sentry
-     * @param int          $level  The minimum logging level at which this handler will be triggered
+     * @param Logger::*    $level  The minimum logging level at which this handler will be triggered
      * @param bool         $bubble Whether the messages that are handled can bubble up the stack or not
      */
-    public function __construct(HubInterface $hub, int $level = Logger::DEBUG, bool $bubble = true)
-    {
+    public function __construct(
+        HubInterface $hub,
+        int $level = Logger::DEBUG,
+        bool $bubble = true
+    ) {
         parent::__construct($level, $bubble);
 
         $this->hub = $hub;
@@ -46,14 +192,10 @@ class SentryHandler extends AbstractProcessingHandler
             return;
         }
 
-        // filter records
-        $records = array_filter(
-            $records,
-            function ($record) {
-                // Keep record that matches the minimum level
-                return $record['level'] >= $this->level;
-            }
-        );
+        // Keep record that matches the minimum level
+        $records = array_filter($records, function ($record) {
+            return $this->isHandling($record);
+        });
 
         if (!$records) {
             return;
@@ -73,7 +215,7 @@ class SentryHandler extends AbstractProcessingHandler
 
         // the other ones are added as a context items
         foreach ($records as $record) {
-            $record              = $this->processRecord($record);
+            $record = $this->processRecord($record);
             $record['formatted'] = $this->getFormatter()->format($record);
 
             $this->breadcrumbsBuffer[] = $record;
@@ -85,38 +227,62 @@ class SentryHandler extends AbstractProcessingHandler
     }
 
     /**
-     * {@inheritdoc}
+     * @param array<string, mixed>|LogRecord $record
      */
-    protected function write(array $record): void
+    protected function doWrite($record): void
     {
-        $sentryEvent = [
-            'level'   => $sentryLevel = $this->getSeverityFromLevel($record['level']),
-            'message' => (new LineFormatter('%channel%.%level_name%: %message%'))->format($record),
-        ];
+        $event = SentryEvent::createEvent();
+        $event->setLevel($this->getSeverityFromLevel((int) $record['level']));
+        $event->setMessage((new LineFormatter('%channel%.%level_name%: %message%'))->format($record));
+        $event->setLogger(sprintf('monolog.%s', $record['channel']));
+
+        $hint = new EventHint();
 
         if (isset($record['context']['exception']) && $record['context']['exception'] instanceof \Throwable) {
-            $sentryEvent['exception'] = $record['context']['exception'];
+            $hint->exception = $record['context']['exception'];
         }
 
-        $this->hub->withScope(function (Scope $scope) use ($record, $sentryEvent, $sentryLevel): void {
-            $scope->setLevel($sentryLevel);
-            $scope->setExtra('monolog.formatted', $record['formatted'] ?? '');
+        $this->hub->withScope(
+            function (Scope $scope) use ($record, $event, $hint): void {
+                $scope->setExtra('monolog.channel', $record['channel']);
+                $scope->setExtra('monolog.formatted', $record['formatted'] ?? '');
+                $scope->setExtra('monolog.level', $record['level_name']);
 
-            foreach ($this->breadcrumbsBuffer as $breadcrumbRecord) {
-                $scope->addBreadcrumb(new Breadcrumb(
-                    $this->getBreadcrumbLevelFromLevel($breadcrumbRecord['level']),
-                    $this->getBreadcrumbTypeFromLevel($breadcrumbRecord['level']),
-                    $breadcrumbRecord['channel'] ?? 'N/A',
-                    $breadcrumbRecord['formatted'] ?? 'N/A'
-                ));
-            }
+                foreach ($this->breadcrumbsBuffer as $breadcrumbRecord) {
+                    $context = array_merge($breadcrumbRecord['context'], $breadcrumbRecord['extra']);
+                    unset($context['exception']);
 
-            $this->processScope($scope, $record, $sentryEvent);
+                    $scope->addBreadcrumb(
+                        new Breadcrumb(
+                            $this->getBreadcrumbLevelFromLevel((int) $breadcrumbRecord['level']),
+                            $this->getBreadcrumbTypeFromLevel((int) $breadcrumbRecord['level']),
+                            (string) $breadcrumbRecord['channel'] ?: 'N/A',
+                            (string) $breadcrumbRecord['message'] ?: 'N/A',
+                            $context
+                        )
+                    );
+                }
 
-            $this->hub->captureEvent($sentryEvent);
-        });
+                $this->processScope($scope, $record, $event);
+
+                $this->hub->captureEvent($event, $hint);
+            });
 
         $this->afterWrite();
+    }
+
+    /**
+     * Translates the Monolog level into the Sentry breadcrumb type.
+     *
+     * @param int $level The Monolog log level
+     */
+    private function getBreadcrumbTypeFromLevel(int $level): string
+    {
+        if ($level >= Logger::ERROR) {
+            return Breadcrumb::TYPE_ERROR;
+        }
+
+        return Breadcrumb::TYPE_DEFAULT;
     }
 
     /**
@@ -125,11 +291,11 @@ class SentryHandler extends AbstractProcessingHandler
      * This method is called when Sentry event is captured by the handler.
      * Override it if you want to add custom data to Sentry $scope.
      *
-     * @param Scope $scope       Sentry scope where you can add custom data
-     * @param array $record      Current monolog record
-     * @param array $sentryEvent Current sentry event that will be captured
+     * @param Scope                          $scope       Sentry scope where you can add custom data
+     * @param array<string, mixed>|LogRecord $record      Current monolog record
+     * @param SentryEvent                    $sentryEvent Current sentry event that will be captured
      */
-    protected function processScope(Scope $scope, array $record, array $sentryEvent): void
+    protected function processScope(Scope $scope, $record, SentryEvent $sentryEvent): void
     {
     }
 
@@ -145,66 +311,10 @@ class SentryHandler extends AbstractProcessingHandler
     {
         $client = $this->hub->getClient();
 
-        if ($client instanceof FlushableClientInterface) {
-            $client->flush();
-        }
-    }
-
-    /**
-     * Translates the Monolog level into the Sentry severity.
-     *
-     * @param int $level The Monolog log level
-     */
-    private function getSeverityFromLevel(int $level): Severity
-    {
-        switch ($level) {
-            case Logger::DEBUG:
-                return Severity::debug();
-            case Logger::INFO:
-            case Logger::NOTICE:
-                return Severity::info();
-            case Logger::WARNING:
-                return Severity::warning();
-            case Logger::ERROR:
-                return Severity::error();
-            default:
-                return Severity::fatal();
-        }
-    }
-
-    /**
-     * Translates the Monolog level into the Sentry breadcrumb level.
-     *
-     * @param int $level The Monolog log level
-     */
-    private function getBreadcrumbLevelFromLevel(int $level): string
-    {
-        switch ($level) {
-            case Logger::DEBUG:
-                return Breadcrumb::LEVEL_DEBUG;
-            case Logger::INFO:
-            case Logger::NOTICE:
-                return Breadcrumb::LEVEL_INFO;
-            case Logger::WARNING:
-                return Breadcrumb::LEVEL_WARNING;
-            case Logger::ERROR:
-                return Breadcrumb::LEVEL_ERROR;
-            default:
-                return Breadcrumb::LEVEL_FATAL;
-        }
-    }
-
-    /**
-     * Translates the Monolog level into the Sentry breadcrumb type.
-     *
-     * @param int $level The Monolog log level
-     */
-    private function getBreadcrumbTypeFromLevel(int $level): string
-    {
-        if ($level >= Logger::ERROR) {
-            return Breadcrumb::TYPE_ERROR;
+        if ($client === null) {
+            return;
         }
 
-        return Breadcrumb::TYPE_DEFAULT;
+        $client->flush();
     }
 }
