@@ -9,43 +9,50 @@ use Coduo\PHPMatcher\PHPUnit\PHPMatcherAssertions;
 use GuzzleHttp\Promise\FulfilledPromise;
 use GuzzleHttp\Promise\PromiseInterface;
 use Monolog\Formatter\LineFormatter;
+use Monolog\Level;
 use Monolog\Logger;
+use Monolog\LogRecord;
 use PHPUnit\Framework\TestCase;
+use Sentry\Breadcrumb;
 use Sentry\ClientBuilder;
 use Sentry\Event;
+use Sentry\Integration\EnvironmentIntegration;
 use Sentry\Options;
+use Sentry\Response;
+use Sentry\ResponseStatus;
+use Sentry\SentrySdk;
 use Sentry\Severity;
-use Sentry\State\Hub;
+use Sentry\State\HubInterface;
 use Sentry\State\Scope;
-use Sentry\Transport\ClosableTransportInterface;
-use Sentry\Transport\NullTransport;
 use Sentry\Transport\TransportFactoryInterface;
 use Sentry\Transport\TransportInterface;
 
-class SentryHandlerTest extends TestCase
+final class SentryHandlerTest extends TestCase
 {
     use PHPMatcherAssertions;
 
-    /**
-     * @var Hub
-     */
-    private $hub;
-
-    /**
-     * @var SpyTransport
-     */
-    private $transport;
+    private HubInterface $hub;
+    private SpyTransport $transport;
 
     protected function setUp(): void
     {
         $this->transport = new SpyTransport();
 
-        $clientBuilder = ClientBuilder::create(['default_integrations' => false]);
+        $clientBuilder = ClientBuilder::create(
+            [
+                'default_integrations' => false,
+                'integrations' => [
+                    // In order to get OS and runtime context we must enable this
+                    new EnvironmentIntegration(),
+                ],
+            ]
+        );
         $clientBuilder->setTransportFactory(new FakeTransportFactory($this->transport));
 
         $client = $clientBuilder->getClient();
 
-        $this->hub = new Hub($client);
+        $this->hub = SentrySdk::getCurrentHub();
+        $this->hub->bindClient($client);
     }
 
     protected function tearDown(): void
@@ -53,18 +60,47 @@ class SentryHandlerTest extends TestCase
         unset($this->hub, $this->transport);
     }
 
+    /**
+     * @return array<string, mixed>|LogRecord
+     */
+    private function record(array $record)
+    {
+        if (Logger::API < 3) { // @phpstan-ignore-line - Comparison operation "<" between 3 and 3 is always false.
+            return $record;
+        }
+
+        return new LogRecord(
+            new \DateTimeImmutable(),
+            $record['channel'],
+            Level::from($record['level']),
+            $record['message'],
+            $record['context'],
+            $record['extra']
+        );
+    }
+
+    /**
+     * @return array<array<string, mixed>|LogRecord>
+     */
+    private function records(array $records): array
+    {
+        return array_map(function ($record) {
+            return $this->record($record);
+        }, $records);
+    }
+
     public function testHandle(): void
     {
         $handler = $this->createSentryHandler();
 
-        $record = [
-            'message'    => 'My info message',
-            'context'    => [],
-            'level'      => Logger::INFO,
+        $record = $this->record([
+            'message' => 'My info message',
+            'context' => [],
+            'level' => Logger::INFO,
             'level_name' => Logger::getLevelName(Logger::INFO),
-            'channel'    => 'app',
-            'extra'      => [],
-        ];
+            'channel' => 'app',
+            'extra' => [],
+        ]);
 
         $handler->handle($record);
 
@@ -80,14 +116,14 @@ class SentryHandlerTest extends TestCase
     {
         $handler = $this->createSentryHandler();
 
-        $record = [
-            'message'    => 'My info message',
-            'context'    => ['exception' => $exception = new \LogicException('Test logic exception')],
-            'level'      => Logger::INFO,
+        $record = $this->record([
+            'message' => 'My info message',
+            'context' => ['exception' => $exception = new \LogicException('Test logic exception')],
+            'level' => Logger::INFO,
             'level_name' => Logger::getLevelName(Logger::INFO),
-            'channel'    => 'app',
-            'extra'      => [],
-        ];
+            'channel' => 'app',
+            'extra' => [],
+        ]);
 
         $handler->handle($record);
 
@@ -113,72 +149,72 @@ class SentryHandlerTest extends TestCase
     {
         $handler = $this->createSentryHandler();
 
-        $records = [
+        $records = $this->records([
             [
-                'message'    => 'Info message',
-                'context'    => ['exception' => new \LogicException()],
-                'level'      => $level = Logger::INFO,
+                'message' => 'Info message',
+                'context' => ['exception' => new \LogicException()],
+                'level' => $level = Logger::INFO,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'chan-info',
-                'extra'      => ['extra-info'],
+                'channel' => 'chan-info',
+                'extra' => ['extra-info'],
             ],
             [
-                'message'    => 'Error Message',
-                'context'    => [],
-                'level'      => $level = Logger::ERROR,
+                'message' => 'Error Message',
+                'context' => [],
+                'level' => $level = Logger::ERROR,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'chan-error',
-                'extra'      => ['extra-error'],
+                'channel' => 'chan-error',
+                'extra' => ['extra-error'],
             ],
             [
-                'message'    => 'Debug message',
-                'context'    => [],
-                'level'      => $level = Logger::DEBUG,
+                'message' => 'Debug message',
+                'context' => [],
+                'level' => $level = Logger::DEBUG,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'chan-debug',
-                'extra'      => ['extra-debug'],
+                'channel' => 'chan-debug',
+                'extra' => ['extra-debug'],
             ],
             [
-                'message'    => 'Emergency message',
-                'context'    => [],
-                'level'      => $level = Logger::EMERGENCY,
+                'message' => 'Emergency message',
+                'context' => [],
+                'level' => $level = Logger::EMERGENCY,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'chan-emerg',
-                'extra'      => ['extra-emerg'],
+                'channel' => 'chan-emerg',
+                'extra' => ['extra-emerg'],
             ],
             [
-                'message'    => 'Warning message',
-                'context'    => [],
-                'level'      => $level = Logger::WARNING,
+                'message' => 'Warning message',
+                'context' => [],
+                'level' => $level = Logger::WARNING,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'chan-warn',
-                'extra'      => ['extra-warn'],
+                'channel' => 'chan-warn',
+                'extra' => ['extra-warn'],
             ],
             [
-                'message'    => 'Notice message',
-                'context'    => [],
-                'level'      => $level = Logger::NOTICE,
+                'message' => 'Notice message',
+                'context' => [],
+                'level' => $level = Logger::NOTICE,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'chan-notice',
-                'extra'      => ['extra-notice'],
+                'channel' => 'chan-notice',
+                'extra' => ['extra-notice'],
             ],
             [
-                'message'    => 'Alert message',
-                'context'    => [],
-                'level'      => $level = Logger::ALERT,
+                'message' => 'Alert message',
+                'context' => [],
+                'level' => $level = Logger::ALERT,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'chan-alert',
-                'extra'      => ['extra-alert'],
+                'channel' => 'chan-alert',
+                'extra' => ['extra-alert'],
             ],
             [
-                'message'    => 'Critical message',
-                'context'    => ['exception' => new \LogicException()],
-                'level'      => $level = Logger::CRITICAL,
+                'message' => 'Critical message',
+                'context' => ['exception' => new \LogicException()],
+                'level' => $level = Logger::CRITICAL,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'chan-critical',
-                'extra'      => ['extra-critical'],
+                'channel' => 'chan-critical',
+                'extra' => ['extra-critical'],
             ],
-        ];
+        ]);
 
         $handler->handleBatch($records);
 
@@ -190,68 +226,72 @@ class SentryHandlerTest extends TestCase
             null,
             [
                 [
-                    'type'      => 'default',
-                    'category'  => 'chan-info',
-                    'level'     => 'info',
-                    'message'   => 'chan-info.INFO: Info message ["extra-info"]',
+                    'type' => 'default',
+                    'category' => 'chan-info',
+                    'level' => 'info',
+                    'message' => 'Info message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => [
+                        'extra-info',
+                    ],
                 ],
                 [
-                    'type'      => 'error',
-                    'category'  => 'chan-error',
-                    'level'     => 'error',
-                    'message'   => 'chan-error.ERROR: Error Message ["extra-error"]',
+                    'type' => 'error',
+                    'category' => 'chan-error',
+                    'level' => 'error',
+                    'message' => 'Error Message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => ['extra-error'],
                 ],
                 [
-                    'type'      => 'default',
-                    'category'  => 'chan-debug',
-                    'level'     => 'debug',
-                    'message'   => 'chan-debug.DEBUG: Debug message ["extra-debug"]',
+                    'type' => 'default',
+                    'category' => 'chan-debug',
+                    'level' => 'debug',
+                    'message' => 'Debug message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => ['extra-debug'],
                 ],
                 [
-                    'type'      => 'error',
-                    'category'  => 'chan-emerg',
-                    'level'     => 'fatal',
-                    'message'   => 'chan-emerg.EMERGENCY: Emergency message ["extra-emerg"]',
+                    'type' => 'error',
+                    'category' => 'chan-emerg',
+                    'level' => 'fatal',
+                    'message' => 'Emergency message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => ['extra-emerg'],
                 ],
                 [
-                    'type'      => 'default',
-                    'category'  => 'chan-warn',
-                    'level'     => 'warning',
-                    'message'   => 'chan-warn.WARNING: Warning message ["extra-warn"]',
+                    'type' => 'default',
+                    'category' => 'chan-warn',
+                    'level' => 'warning',
+                    'message' => 'Warning message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => ['extra-warn'],
                 ],
                 [
-                    'type'      => 'default',
-                    'category'  => 'chan-notice',
-                    'level'     => 'info',
-                    'message'   => 'chan-notice.NOTICE: Notice message ["extra-notice"]',
+                    'type' => 'default',
+                    'category' => 'chan-notice',
+                    'level' => 'info',
+                    'message' => 'Notice message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => ['extra-notice'],
                 ],
                 [
-                    'type'      => 'error',
-                    'category'  => 'chan-alert',
-                    'level'     => 'fatal',
-                    'message'   => 'chan-alert.ALERT: Alert message ["extra-alert"]',
+                    'type' => 'error',
+                    'category' => 'chan-alert',
+                    'level' => 'fatal',
+                    'message' => 'Alert message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => ['extra-alert'],
                 ],
                 [
-                    'type'      => 'error',
-                    'category'  => 'chan-critical',
-                    'level'     => 'fatal',
-                    'message'   => 'chan-critical.CRITICAL: Critical message ["extra-critical"]',
+                    'type' => 'error',
+                    'category' => 'chan-critical',
+                    'level' => 'fatal',
+                    'message' => 'Critical message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => [
+                        'extra-critical',
+                    ],
                 ],
             ]
         );
@@ -261,56 +301,56 @@ class SentryHandlerTest extends TestCase
     {
         $handler = $this->createSentryHandler(Logger::WARNING);
 
-        $records = [
+        $records = $this->records([
             [
-                'message'    => 'Info message',
-                'context'    => ['exception' => new \LogicException()],
-                'level'      => $level = Logger::INFO,
+                'message' => 'Info message',
+                'context' => ['exception' => new \LogicException()],
+                'level' => $level = Logger::INFO,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'test',
-                'extra'      => [],
+                'channel' => 'test',
+                'extra' => [],
             ],
             [
-                'message'    => 'Error Message',
-                'context'    => [],
-                'level'      => $level = Logger::ERROR,
+                'message' => 'Error Message',
+                'context' => [],
+                'level' => $level = Logger::ERROR,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'test',
-                'extra'      => [],
+                'channel' => 'test',
+                'extra' => [],
             ],
             [
-                'message'    => 'Debug message',
-                'context'    => [],
-                'level'      => $level = Logger::DEBUG,
+                'message' => 'Debug message',
+                'context' => [],
+                'level' => $level = Logger::DEBUG,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'test',
-                'extra'      => [],
+                'channel' => 'test',
+                'extra' => [],
             ],
             [
-                'message'    => 'Warning message',
-                'context'    => [],
-                'level'      => $level = Logger::WARNING,
+                'message' => 'Warning message',
+                'context' => [],
+                'level' => $level = Logger::WARNING,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'test',
-                'extra'      => [],
+                'channel' => 'test',
+                'extra' => [],
             ],
             [
-                'message'    => 'Notice message',
-                'context'    => [],
-                'level'      => $level = Logger::NOTICE,
+                'message' => 'Notice message',
+                'context' => [],
+                'level' => $level = Logger::NOTICE,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'test',
-                'extra'      => [],
+                'channel' => 'test',
+                'extra' => [],
             ],
             [
-                'message'    => 'Critical message',
-                'context'    => ['exception' => $exception = new \LogicException('Exception of critical level')],
-                'level'      => $level = Logger::CRITICAL,
+                'message' => 'Critical message',
+                'context' => ['exception' => $exception = new \LogicException('Exception of critical level')],
+                'level' => $level = Logger::CRITICAL,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'test',
-                'extra'      => [],
+                'channel' => 'test',
+                'extra' => [],
             ],
-        ];
+        ]);
 
         $handler->handleBatch($records);
 
@@ -322,28 +362,28 @@ class SentryHandlerTest extends TestCase
             $exception,
             [
                 [
-                    'type'      => 'error',
-                    'category'  => 'test',
-                    'level'     => 'error',
-                    'message'   => 'test.ERROR: Error Message []',
+                    'type' => 'error',
+                    'category' => 'test',
+                    'level' => 'error',
+                    'message' => 'Error Message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => [],
                 ],
                 [
-                    'type'      => 'default',
-                    'category'  => 'test',
-                    'level'     => 'warning',
-                    'message'   => 'test.WARNING: Warning message []',
+                    'type' => 'default',
+                    'category' => 'test',
+                    'level' => 'warning',
+                    'message' => 'Warning message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => [],
                 ],
                 [
-                    'type'      => 'error',
-                    'category'  => 'test',
-                    'level'     => 'fatal',
-                    'message'   => 'test.CRITICAL: Critical message []',
+                    'type' => 'error',
+                    'category' => 'test',
+                    'level' => 'fatal',
+                    'message' => 'Critical message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => [],
                 ],
             ]
         );
@@ -353,16 +393,16 @@ class SentryHandlerTest extends TestCase
     {
         $handler = $this->createSentryHandler();
 
-        $records = [
+        $records = $this->records([
             [
-                'message'    => 'Info message',
-                'context'    => [],
-                'level'      => $level = Logger::INFO,
+                'message' => 'Info message',
+                'context' => [],
+                'level' => $level = Logger::INFO,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'test',
-                'extra'      => [],
+                'channel' => 'test',
+                'extra' => [],
             ],
-        ];
+        ]);
 
         $handler->handleBatch($records);
 
@@ -381,12 +421,12 @@ class SentryHandlerTest extends TestCase
             null,
             [
                 [
-                    'type'      => 'default',
-                    'category'  => 'test',
-                    'level'     => 'info',
-                    'message'   => 'test.INFO: Info message []',
+                    'type' => 'default',
+                    'category' => 'test',
+                    'level' => 'info',
+                    'message' => 'Info message',
                     'timestamp' => '@double@',
-                    'data'      => [],
+                    'data' => [],
                 ],
             ]
         );
@@ -396,24 +436,24 @@ class SentryHandlerTest extends TestCase
     {
         $handler = $this->createSentryHandler(Logger::EMERGENCY);
 
-        $records = [
+        $records = $this->records([
             [
-                'message'    => 'Info message',
-                'context'    => ['exception' => new \LogicException()],
-                'level'      => $level = Logger::INFO,
+                'message' => 'Info message',
+                'context' => ['exception' => new \LogicException()],
+                'level' => $level = Logger::INFO,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'test',
-                'extra'      => [],
+                'channel' => 'test',
+                'extra' => [],
             ],
             [
-                'message'    => 'Critical message',
-                'context'    => ['exception' => $exception = new \LogicException('Exception of critical level')],
-                'level'      => $level = Logger::CRITICAL,
+                'message' => 'Critical message',
+                'context' => ['exception' => $exception = new \LogicException('Exception of critical level')],
+                'level' => $level = Logger::CRITICAL,
                 'level_name' => Logger::getLevelName($level),
-                'channel'    => 'test',
-                'extra'      => [],
+                'channel' => 'test',
+                'extra' => [],
             ],
-        ];
+        ]);
 
         $handler->handleBatch($records);
 
@@ -423,61 +463,74 @@ class SentryHandlerTest extends TestCase
 
     private function assertCapturedEvent(Severity $severity, string $message, array $extra, \Exception $exception = null, array $breadcrumbs = []): void
     {
-        $event = $this->transport->getSpiedEventsAsArray();
+        $event = $this->transport->getSpiedEvent();
 
         if (null !== $exception) {
-            $this->assertCount(1, $event['exception']['values']);
-            $this->assertSame(\get_class($exception), $event['exception']['values'][0]['type']);
-            $this->assertSame($exception->getMessage(), $event['exception']['values'][0]['value']);
+            $this->assertCount(1, $event->getExceptions());
+            $exceptionDataBag = $event->getExceptions()[0];
+            $this->assertSame(\get_class($exception), $exceptionDataBag->getType());
+            $this->assertSame($exception->getMessage(), $exceptionDataBag->getValue());
         } else {
-            $this->assertArrayNotHasKey('exception', $event);
+            $this->assertCount(0, $event->getExceptions());
         }
 
         $this->assertTrue($this->transport->isFlushed);
-        $this->assertArrayNotHasKey('tags', $event);
-        $this->assertArrayNotHasKey('user', $event);
-        $this->assertSame($message, $event['message']);
-        $this->assertSame((string) $severity, $event['level']);
-        $this->assertMatchesPattern('@string@', $event['event_id']);
-        $this->assertMatchesPattern('@string@', $event['timestamp']);
-        $this->assertMatchesPattern('@string@', $event['platform']);
-        $this->assertMatchesPattern('@string@', $event['server_name']);
-        $this->assertMatchesPattern(['processScope' => 'called'] + $extra, $event['extra']);
+        $this->assertCount(0, $event->getTags());
+        $this->assertNull($event->getUser());
+        $this->assertSame($message, $event->getMessage());
+        $this->assertSame((string) $severity, (string) $event->getLevel());
+        $this->assertMatchesPattern('@string@', (string) $event->getId());
+        $this->assertMatchesPattern('@string@', (string) $event->getTimestamp());
+        $this->assertMatchesPattern('@string@', $event->getServerName());
+        $this->assertMatchesPattern([
+                'processScope' => 'called',
+                'monolog.channel' => '@string@',
+                'monolog.level' => '@string@',
+            ] + $extra,
+            $event->getExtra(),
+        );
 
         if ($breadcrumbs) {
             $this->assertMatchesPattern(
                 json_encode($breadcrumbs),
-                json_encode($event['breadcrumbs']['values'])
+                json_encode(
+                    array_map(
+                        static function (Breadcrumb $breadcrumb) {
+                            return [
+                                'type' => $breadcrumb->getType(),
+                                'category' => $breadcrumb->getCategory(),
+                                'level' => $breadcrumb->getLevel(),
+                                'message' => $breadcrumb->getMessage(),
+                                'timestamp' => (float) $breadcrumb->getTimestamp(),
+                                'data' => $breadcrumb->getMetadata(),
+                            ];
+                        },
+                        $event->getBreadcrumbs()
+                    )
+                )
             );
         } else {
-            $this->assertArrayNotHasKey('breadcrumbs', $event);
+            $this->assertCount(0, $event->getBreadcrumbs());
         }
 
-        $this->assertMatchesPattern(
-            [
-                'name'    => 'sentry.php',
-                'version' => '@string@',
-            ],
-            $event['sdk']
-        );
+        $this->assertSame('sentry.php', $event->getSdkIdentifier());
+        $this->assertMatchesPattern('@string@', $event->getSdkVersion());
 
-        $this->assertMatchesPattern(
-            [
-                'os' => [
-                    'name'           => '@string@',
-                    'version'        => '@string@',
-                    'build'          => '@string@',
-                    'kernel_version' => '@string@',
-                ],
-                'runtime' => [
-                    'name'    => 'php',
-                    'version' => '@string@',
-                ],
-            ],
-            $event['contexts']
-        );
+        if (null !== ($os = $event->getOsContext())) {
+            $this->assertMatchesPattern('@string@', $os->getName());
+            $this->assertMatchesPattern('@string@', $os->getVersion());
+            $this->assertMatchesPattern('@string@', $os->getBuild());
+            $this->assertMatchesPattern('@string@', $os->getKernelVersion());
+        }
+        if (null !== ($runtime = $event->getRuntimeContext())) {
+            $this->assertMatchesPattern('php', $runtime->getName());
+            $this->assertMatchesPattern('@string@', $runtime->getVersion());
+        }
     }
 
+    /**
+     * @param Logger::* $level
+     */
     private function createSentryHandler(int $level = null): SpySentryHandler
     {
         if (null === $level) {
@@ -500,7 +553,7 @@ class SpySentryHandler extends SentryHandler
     public $afterWriteCalled = false;
 
     /** {@inheritdoc} */
-    protected function processScope(Scope $scope, array $record, array $sentryEvent): void
+    protected function processScope(Scope $scope, $record, Event $sentryEvent): void
     {
         $scope->setExtra('processScope', 'called');
     }
@@ -519,7 +572,7 @@ class SpySentryHandler extends SentryHandler
     }
 }
 
-class SpyTransport extends NullTransport implements ClosableTransportInterface
+class SpyTransport implements TransportInterface
 {
     /**
      * @var Event|null
@@ -531,29 +584,26 @@ class SpyTransport extends NullTransport implements ClosableTransportInterface
      */
     public $isFlushed = false;
 
-    public function send(Event $event): ?string
+    public function send(Event $event): PromiseInterface
     {
         $this->spiedEvent = $event;
 
-        return parent::send($event);
+        return new FulfilledPromise(new Response(ResponseStatus::skipped(), $event));
     }
 
     public function resetSpy(): void
     {
         $this->spiedEvent = null;
-        $this->isFlushed  = false;
+        $this->isFlushed = false;
     }
 
-    /**
-     * @return array<string, mixed>
-     */
-    public function getSpiedEventsAsArray(): array
+    public function getSpiedEvent(): Event
     {
         if (null === $this->spiedEvent) {
             throw new \RuntimeException('No spied scope');
         }
 
-        return $this->spiedEvent->toArray();
+        return $this->spiedEvent;
     }
 
     public function close(?int $timeout = null): PromiseInterface
